se añadio el controller y la funcion para listar categorias

+++ src/main/java/com/piero/backend/chat/app/platos/dto/categoria/CategoriaDTOResponse.java
@@ -1,8 +1,9 @@
 package com.piero.backend.chat.app.platos.dto.categoria;
 
+
 public record CategoriaDTOResponse (
-    Integer id,
-    String nombre,
-    String descripcion
+        Short id,
+        String nombre,
+        String descripcion
 ){
 }
